clean up header files
